Release version 1.1.0 - performance improvements
